Optimize delay resizing (#3)
Gets rid of the old algorithm that would resize the delay line itself. Thanks to this, there's no more aggressive clicking noise when the length of the delay line changes.
Instead, the delay line has now a fixed length and the samples get time stretched (resized with linear interpolation) before feeding them to the delay line and the output of the delay line is resized back to the original size.
The above results in the same pitch shift effect as the delay "length" changes as before, but the clicking noise is eliminated.

Unfortunately some form of "digital sounding" noise still remains at shorter delay lengths and it results either from bad interpolation of the last element of the buffer, or noise introduced by linear interpolation that's of rather poor quality.